backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["swadhinbiswas <swadhinbiswas.cse@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.7"}
openai = "^1.59.9"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.37"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
bcrypt = "^4.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
python-telegram-bot = "^21.10"
google-genai = "^1.0.0"
qdrant-client = "^1.13.2"
google-generativeai = "^0.8.4"
pillow = "^11.1.0"
aiohttp = "^3.11.12"
fake-useragent = "^2.0.3"
selenium = "^4.28.1"
beautifulsoup4 = "^4.13.3"
webdriver-manager = "^4.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
